GUI semi done
There is now a Client GUI.
If you start the ThreadedServer, and then start the SocketGUI.
Then it should be able to connect to the Server with the connect button.

You can then type in commands in the TextArea, and press send.
The Server will then give you an output.

The only thing that I can't figure out...
Is the while (true) loop. For some reason, it freezes the window.. I don't know why, I won't know how to avoid it.